fixup: 64-bit warning

+++ source/slang/slang-compiler.cpp
@@ -1014,7 +1014,7 @@ namespace Slang
             SourceView* sourceView = sourceManager->createSourceView(sourceFile, &filePathInfo);
 
             *ppData = sourceView->getContent().begin();
-            *pBytes = sourceView->getContentSize();
+            *pBytes = (UINT) sourceView->getContentSize();
 
             return S_OK;
         }
